fn stmt no capture
now you need to assign a fn expr to a constant. made it more like rust which is funny cause that always annoyed me so i might have to change it back. but its convenient for gradual resolve to not worry about rt captures and it makes it more obvious that closures cant be added to an overload set